Moving tags and sites update and eventbus subscriber.
